requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
kagglehub>=0.2.0
SQLAlchemy>=2.0.0
alembic>=1.12.0

# API and async
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0

# Text processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
nltk>=3.8.0

# CLI and output
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Configuration
PyYAML>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Machine learning (for sampling)
scikit-learn>=1.3.0